reorganise package into team, parse, serialize
